Shutter Ace reaches its main menu: EVALUATE suspends, errors cross fibers, and what CreateGlobalObjects needed
EVALUATE() runs its text as an inline frame of the calling routine, as &expr does, so an
expression that calls a method can wait on the host (CodeMine's cmEvent.Subscribe).

A method runs as a fiber of its own; the scheduler now tells the VM which fiber it runs for, so
an error nothing in the method handles reaches a TRY in the caller (measured: nearer than ON
ERROR, not nearer than the object's Error method). Calls past VFP's level 127 (126 for a method)
raise error 103, which also keeps runaway recursion from overflowing the JS stack and poisoning
the VM.

Also, each measured in vfp9.exe: DODEFAULT() from a library object added into another object;
VARTYPE() of an undefined name; TYPE() of o.Parent; SKIP IN o.cWorkarea; DIMENSION &name[n];
SET MEMOWIDTH to 8192; SET CENTURY TO (expr) ROLLOVER (expr); SET RELATION OFF INTO (x) IN (y);
a macro standing for several arguments; CURSORGETPROP SourceType and Database; GETFLDSTATE;
numeric index on an empty table; FULLPATH and OPEN DATABASE along SET PATH; CREATE DATABASE
data\x; MLINE offsets and _MLINE; SYS(16, n); PROGRAM() and LINENO() inside ON ERROR; SCATTER
NAME; array functions filling a property; AddObject of a library class; Load only for visual
objects; class property expressions with header constants, NULL and THIS; Exception.Procedure
and Error cMethod spelled as the product spells them.

Fixes found on the way: carrying on after an error in a macro, EVALUATE or ON ERROR handler
looped on a stack underflow; the memory API dropped appended records; a host read's HostError
is raised where the read was.

+++ crates/foxvm/src/value.rs
@@ -452,7 +452,7 @@ pub struct Settings {
     pub unique: bool,
     /// SET MEMOWIDTH: the width a memo is taken to wrap at, which is what decides the line a
     /// character of it sits on. 50 is VFP default.
-    pub memowidth: u8,
+    pub memowidth: u16,
     /// SET HEADINGS ON (the default): LIST and DISPLAY of records write a row of field names
     /// above them. It says nothing about LIST STRUCTURE, which always names its columns.
     pub headings: bool,
